Allow for annotating violations in PRs
This commit adds a new input to allow for hooking into the GitHub annotations API.

This allows for flagging any violations flagged by `xmllint` in the code view of PRs.
It also allows for the violations to be rendered in the "Annotations" fold-out in GH Actions job pages.

By default, the `show-in-pr` flag is set to `true`.

Includes tests.

Ref: https://github.com/korelstar/xmllint-problem-matcher

---

Please note that even though GH Action `input` meta data is supposed to allow for a "default" value, this doesn't appear to work correctly.
Might be that it doesn't work for composite actions, might be that only string values are supported (though I tested that too and still had no luck), but either way, the fact that we can't rely on a meaningful default value is problematic.

In other words, to support this toggle flag properly, we can't use an `if: ${{ input.show-in-pr }}` check, as it won't behave as expected.
Instead we need to explicitly check against a `false` value to safeguard the default behaviour (annotations = on) is respected.