Allow for validating against a locally available XSD file
This commit add a new input `xsd-file` which takes a path to a file.
When this file is available and identifies as a non-empty `xsd` file, the XML file(s) provided via the `pattern` input will be validated against the provided XSD file.

Includes tests.

Props to binarykitten for helping to debug the bash scripting.

Co-authored-by: Kathryn Reeve <67553+BinaryKitten@users.noreply.github.com>