QA/Tests: add minimal GH Pages set-up
In the next commit, an option to use a remote XSD file to validate against will be introduced.

To allow for testing this, we need to ensure there are some downloadable XSD files available at publicly accessible URLs.

This commit sets up these files in a `docs` directory, which will earmarked to be auto-deployed to a plain GH Pages website.
This should allow for the upcoming tests to download these XSD files during the action run.

By having these files deployed to a GH Pages website, the project doesn't need to have an external hosting dependency and remains self-contained.