Allow for validating against a remote XSD file
This commit add a new input `xsd-url` which takes a URL for the XSD file.
The workflow will attempt to download the XSD file and if succesful and the file identifies as a non-empty `xsd` file, the XML file(s) provided via the `pattern` input will be validated against the provided (remote) XSD file.

If both the `xsd-file` as well as the `xsd-url` inputs are provided, the local `xsd-file` will always take precedence over the (remote) `xsd-url`.

Includes tests.