Add a "debug" mode with more verbose output
This commit adds a new input to allow for debugging the action runner by showing more verbose output, including showing the XMLLint version used and showing what files will be matched by a glob pattern.

By default, the `debug` flag is set to `false` (= "off").

Includes tests.

---

Please note that even though GH Action `input` meta data is supposed to allow for a "default" value, this doesn't appear to work correctly.
Might be that it doesn't work for composite actions, might be that only string values are supported (though I tested that too and still had no luck), but either way, the fact that we can't rely on a meaningful default value is problematic.

In other words, to support this toggle flag properly, we can't use an `if: ${{ input.debug }}` check, as it won't behave as expected.
Instead we need to explicitly check against a `true` value to safeguard that the debug mode is only "on" when explicitly requested.

Also take note that, as done in other commits, when steps now become multi-line to handle the new input option correctly, a comment is added on the first line of the multi-line instructions to make the output more readable/understandable.